Created a resolver to fetch trends
I though there was something wrong with the way
angular fetched the data that was messing with flickity
(it wasn't displaying alright). Big mistake, I had to initialize my
Flickity Object in the ngAfterViewInit lifeCycle hook.

I might remove this resolver in the future, and just fetch the data on
init instaed.